python/ai-todo/backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
google-generativeai
scikit-learn
nltk
sentence-transformers
apscheduler
pywebpush
numpy
requests